+++ transport_indirect.h
@@ -38,8 +38,8 @@ class TransportIndirect : public Transport {
 public:
 	TransportIndirect(FLContext *handle, const char *conduit);
 	void sendMessage(
-		const uint8 *cmdData, size_t cmdLength = 1,
-		uint8 *recvBuf = 0, size_t recvLength = 0
+		const uint8 *cmdData, uint32 cmdLength = 1,
+		uint8 *recvBuf = 0, uint32 recvLength = 0
 	) const;
 };
